pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "icecream-simulator"
version = "0.3.0"
description = "Modular simulation framework for ice cream production and waste-to-plastic conversion"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
]
dashboard = [
    "streamlit>=1.28",
]
config = [
    "pyyaml>=6.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
